added sparse matrix header
